deploy/frontend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
google-auth>=2.29.0
firebase-admin>=6.5.0

